Change dn121 model_features from 1024 to 2048
After trying to use dn121 and getting an error when doing precompute = True I tracked it down to this as the issue.  The activation generated is 2048 where the bcolz empty file that is generated is 1024 giving a `ValueError: array trailing dimensions do not match with self` error.  The change I made just makes the initial file 2048 wide instead of 1024.  This fixes the issue, but I'm not 100% if the 1024 is correct and it is actually the model that shouldn't be generating 2048 wide activations.  If that is the case, I need to dig into that side instead.