Adds use_clr_beta
Matches the 1cycle policy of Leslie. Typical use is use_clr_beta = (div,pct,max_mom,min_mom)
The two last args can be skipped if you don't want to use cyclical momentum.
The first one is the amount to divide the passed learning rate to get the minimum learning rate, the second one the part of the cycle (in percent) that will be devoted to the LR annealing after the traingular cycle.