Bug in validate
With a pdb.set_trace() I realized that in the validate function, x is always a list containing a Tensor with our minibatch. Then the current command batch_cnts.append(len(x)) always add ones instead of the actual size of the minibatch which makes the results false (the last minibatch usually has a smaller size).