Proposition to refactor sgdr
Separates the different phases of training with GroupEpochs. Each phase can have its own parameters, including the optimizer.

For now it's independant of the existing classes to not break anything while we debug but ultimaltey it can be used to do all the existing schedulers.

To use, build a list of EpochGroup (a better name would perhaps be TrainingPhase?) and use learn.fit_opt_sched(epoch_groups).

To change the dataset used (it will bug if there is a change of batch_size) use the additional argument list_data (must be the same length as epoch_groups).