Fixes bug
On_phase_begin always needs to be called **before** passing the optimizer to the stepper (otehrwise the stepper has an optimizer that isn't updated).

Also, changing the optimizer causes to zero all the buffers in it (which is kind of logical) so we lose all the previous moving averages. For now I only create a new one if there is a change of name (so if we keep SGD with momentum for instance we don't reinitialize the moving average of the gradients), but perhaps it could be interesting to 'match' the buffers in some way?