Update pandas-summary dependency
Previously the structured data analysis notebook (course 1, lesson 3) would
fail on the DataFrameSummary command due to there being a change in how Pandas
handled types and pandas-summary. There is a change in pandas-summary, but the
version that released it seems to have gone down [1], so conda doesn't know how
to install it. So we have to hardcode it to the newly released version until
the version changes.

[1]: https://github.com/mouradmourafiq/pandas-summary/issues/14