add missing lesson1-breeds.ipynb notebook.
In the first few DL lessons Jeremy often works with dog breeds notebook (lesson1-breeds.ipynb), but says he can't give it us since the competition is unfinished and he doesn't want us to cheat.

The competition has been finished, so let's make this notebook available to all so it's easier to follow the video lessons.

I found a copy of it in this repo https://github.com/tanakatsu/fastai-course1-v2. I can't tell whether it is identical to Jeremy's copy and how it was acquired.
But it's probably better to have it in any form than none.

I tested it that it runs to a completion as is.

Here are the changes I have applied to lesson1-breeds.ipynb found under the github repo mentioned above:

- cleaned up the corrections that were made due to the API changes made since the video lessons.
- added a few notes where to find data needed for that notebook to run.
- edited the last section to remove the non-working code and clarify that there are 2 methods.
- removed the hardcoded section numbers (jupyter generates those automatically).